# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/login": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "Logs in an existing user.(Remember to signup a user first)",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "You are logged in"
          },
          "400": {
            "description": "Password is required."
          },
          "401": {
            "description": "Unauthorized: Invalid token"
          },
          "500": {
            "description": "Something went wrong with creating JWT token"
          }
        }
      }
    },
    "/signup": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "Signs up a user.",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Signup"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "You created an account."
          },
          "400": {
            "description": "Provided email is already in use."
          },
          "401": {
            "description": "Unauthorized: Invalid token"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Login": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "user@example.com"
        },
        "password": {
          "type": "string",
          "example": "P4ssword"
        }
      },
      "required": [
        "email",
        "password"
      ]
    },
    "Signup": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "User"
        },
        "email": {
          "type": "string",
          "example": "user@example.com"
        },
        "password": {
          "type": "string",
          "example": "P4ssword"
        }
      },
      "required": [
        "name",
        "email",
        "password"
      ]
    },
...
  }
}